+++ packages/logger/vitest.config.ts
@@ -3,14 +3,17 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     globals: true,
+    clearMocks: true,
     coverage: {
+      provider: 'v8',
       reporter: ['text'],
       include: ['src/lib'],
-      all: true,
-      lines: 80,
-      functions: 80,
-      branches: 80,
-      statements: 80,
+      thresholds: {
+        lines: 80,
+        functions: 80,
+        branches: 80,
+        statements: 80,
+      },
     },
   },
 });