+++ packages/express/CHANGELOG.md
@@ -1,16 +1,25 @@
 # @wesp-up/express
 
+## 2.1.0 (2024-08-20)
+
+### Minor Changes
+
+- Added: The ability to skip access logs for specific routes.
+- Added: Default to format success status codes (2xx, and 3xx) in metrics.
+- Changed: Normalizing the request path in metrics uses the URL path rather than the Express route path. This allows for more accurate reporting when Express middleware or parent routes are used. For example, `app.use('*', myMiddleware)` would result in `*` as the path.
+- Changed: Completely removed the deprecated request context from the response locals object.
+
 ## 2.0.1 (2024-08-20)
 
 ### Patch Changes
 
-- Allowed support for any express version after 4.18.2 rather than only 4.18.2 in peer dependencies.
+- Fixed: Allowed support for any express version after 4.18.2 rather than only 4.18.2 in peer dependencies.
 
 ## 2.0.0 (2024-07-28)
 
 ### Major Changes
 
-- Moved context to req.context rather than res.locals.requestContext. Added a request logger. Excluding 404s from metrics. Added a custom 500 error handler.
+- Changed: Moved context to req.context rather than res.locals.requestContext. Added a request logger. Excluding 404s from metrics. Added a custom 500 error handler.
 
 ## 1.0.1 (2023-11-25)